requirements.txt
requests
PyYAML
tqdm
customtkinter
playsound==1.2.2 